fix MSVC warning levels, move _CRT_SECURE_NO_WARNINGS define to sources

+++ CuTest.h
@@ -8,7 +8,7 @@
 
 /* CuString */
 
-char* CuStrAlloc(int size);
+char* CuStrAlloc(size_t size);
 char* CuStrCopy(const char* old);
 
 #define CU_ALLOC(TYPE)		((TYPE*) malloc(sizeof(TYPE)))
